requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
PyAutoGUI==0.9.54
comtypes==1.4.7
pycaw==20240210
protobuf==4.25.3
screen-brightness-control==0.23.0
Pillow==10.4.0
streamlit==1.35.0
